Constructors

InvariantStopping.Schedule — Method

Schedule(::LinRange)

Constructor for Schedule. Creates a schedule representing a trajectory that stops at times given by the argument.

Examples

julia> schedule = Schedule(LinRange(0,10,11))

InvariantStopping.Schedule — Method

Schedule(::LinRange, ::Integer)

Constructor for Schedule. Creates a schedule representing a tree of trajectories with a prescribed branching factor. Examples:

julia> schedule = Schedule(LinRange(0,10,11),2)

InvariantStopping.Sample — Method

Sample(::State, ::Schedule, ::UnderlyingModel)

Constructor for Sample. Starting from the initial State, it samples from the UnderlyingModel according to the Schedule.

Examples

julia> initial_state = State(0.0, 4.5)

julia> schedule = Schedule(LinRange(0,10,5))

julia> underlying_model = GeometricBrownianMotion() # Default parameters

julia> sample = Sample(initial_state, schedule, underlying_model)

State Transition

InvariantStopping.forward_to — Function

forward_to(::State, ::Number, ::UnderlyingModel)

Forwards a (State)(@ref) in time using the provided ()UnderlyingModel](@ref). Returns a (State)(@ref).

Examples

julia> initial_state = State(0.0,1.0)

julia> forward_time = 10.0

julia> underlying_model = GeometricBrownianMotion()

julia> forward_state = forward_to(initial_state, forward_time, underlying_model)