Overview

State Space

InvariantStopping.State — Type

State

Fields

time: Represents concrete time.

coord: An N dimensional coordinate.

Comment

A state is an immutable object, perhaps it should be made mutable.

Comment

We still debate as whether or not to include the time component inside the state object.

Examples

julia> state = State(0.0, (1.0,0.0))

julia> state = State(0.0, (1.0,))

julia> state = State(0.5, 1.0)

julia> time = get_time(state)

julia> coord = get_coord(state)

Stopping Times

InvariantStopping.StoppingPolicy — Type

StoppingPolicy

Abstract type that specifies when to stop.

It is either DeterministicStopping or StoppingTime.

InvariantStopping.DeterministicStopping — Type

DeterministicStopping <: StoppingPolicy

This type represents stopping at a specific concrete time as opposed to following a generic stopping time.

Examples

julia> simple_stopping_rule = DeterministicStopping(5.0)

julia> time = get_time(simple_stopping_rule) # 5.0

InvariantStopping.StoppingTime — Type

StoppingTime <: StoppingPolicy

This abstract type represents a generic stopping time.

Ordering Stopping Times

InvariantStopping.Schedule — Type

Schedule

An abstract type representing partial orders over StoppingPolicy.

InvariantStopping.NodeSchedule — Type

NodeSchedule <: Schedule

This type stores a StoppingPolicy and a collection of Schedule. Each of the later is to be interpreted as StoppingPolicy to happen at a later point in time.

Examples

julia> stopping_policy_1 = DeterministicStopping(3.0)

julia> stopping_policy_2 = DeterministicStopping(4.0)

julia> leaf = LeafSchedule(stopping_policy_2)

julia> node = NodeSchedule(stopping_policy_1, (leaf,leaf))

InvariantStopping.LeafSchedule — Type

LeafSchedule <: Schedule

This type stores a StoppingPolicy. To be interpreted as the last StoppingPolicy to be followed.

Forwarding States

InvariantStopping.UnderlyingModel — Type

UnderlyingModel

An abstract type representing the transition function of a markov process.

InvariantStopping.GeometricBrownianMotion — Type

GeometricBrownianMotion <: UnderlyingModel

A geometric Brownian motion, specified by its rate, standard deviation sigma and dividend.

julia> underlying_model = GeometricBrownianMotion(3.1,2.0,0.0)

Sampling

InvariantStopping.Sample — Type

Sample

Abstract type representing a realization.

!!! Note The Sample object can be thought as a dictionary where the elements of the Schedule object are the keys and where the states are the items. One samples a random dictionary from schedule to state.

InvariantStopping.NodeSample — Type

NodeSample <: Sample

This type stores a State, a Schedule, an UnderlyingModel and a collection of Sample. Each of the later is to be interpreted as happening at a later point in time.

InvariantStopping.LeafSample — Type

Leaf <: Sample

This type stores a State, a Schedule, an UnderlyingModel. To be interpreted as the last piece of sample data along a schedule branch.