InvariantStopping.BrownianMotion — Type

BrownianMotion

InvariantStopping.GeometricBrownianMotion — Type

GeometricBrownianMotion <: UnderlyingModel

A geometric Brownian motion, specified by its rate, standard deviation sigma and dividend.

julia> underlying_model = GeometricBrownianMotion(3.1,2.0,0.0)

InvariantStopping.LoweredSample — Type

LoweredSample

A unit of sampling.

InvariantStopping.LoweredSchedule — Type

LoweredSchedule

A schedule like object that services the need of the sampler. The timeline contains the sampling events (timestamps) from the earliest stopping opportunity of the parent schedule to the last stopping opportunity of the schedule.

InvariantStopping.Sample — Type

Sample

A state sampled according to a given stopping time. It is set to 'nothing' if the stopping time is never realized.

InvariantStopping.Schedule — Type

Schedule

Specifies a directed tree over stopping times.

InvariantStopping.State — Type

State

Contains the coordinates of a given realization at a given time.

InvariantStopping.StoppingOpportunity — Type

StoppingOpportunity

A Stopping Opportunity is a fundamental unit of stopping: it can only stop at one specific time. It consists of

A list of timestamps. The last timestamp in this list specifies when the predicate is evaluated to decide on stopping.

A function that determines whether stopping should occur based on the current time and the states associated with each timestamp in timestamp_list.

InvariantStopping.StoppingTime — Type

StoppingTime

A stopping time stops at the first stopping opportunity for which its predicate returns true.

It consists simply of a list of stopping opportunities.

InvariantStopping.UnderlyingModel — Type

UnderlyingModel

Abstract class for processes that implement the forward method.

InvariantStopping.DeterministicTime — Method

DeterministicTime

InvariantStopping.HittingTime — Method

HittingTime

Evalues the provided predicate at each timestamp in timestamp_list.

InvariantStopping.Star — Method

Star(::Any,::Integer)

Constructs a tree where the branching factor is applied to the first timestamp only.

InvariantStopping.Tree — Method

Tree

Constructs a tree of schedule with provided branching factor.

InvariantStopping.forward — Method

forward

Simulate the process from the given state from now to a later time. Returns the state generated at the later time. Subtype of UnderlyingModel are informally required to implement the forward method.

InvariantStopping.get_all_leaf — Method

get_all_leaf

Returns a list of all leafs from the provided sample onward.

InvariantStopping.get_all_timestamp — Method

get_all_timestamp

InvariantStopping.get_all_trajectory — Method

get_all_trajectory

Returns a list of all trajectories from the provided sampled onward. A trajectory is a list of states from the provided sample onward to a leaf.

InvariantStopping.get_history — Method

get_history

Returns the list of samples starting from the root to the provided sample.

InvariantStopping.get_sample — Method

get_sample

Populates the provided schedule by simulating the underlying process from the provided state onward.

InvariantStopping.lower — Method

lower

Recursively build a lowered schedule from the provided schedule.

InvariantStopping.plot — Method

plot(::Union{Sample,LoweredSample})

1D : Plot a sample with the x-axis representing time and the y-axis representing the first coordinate.

InvariantStopping.plot_lower — Method

plot_lower