Underlying Model

State

InvariantStopping.SimulationState.State — Type

State

Fields

time: Represents concrete time.

julia coord: An N dimensional coordinate.

A state is an object that contains the time and coordinates of a given realization.

StoppingTime

InvariantStopping.Policy.StoppingTime — Type

StoppingTime

InvariantStopping.Policy.StoppingOpportunity — Type

StoppingOpportunity

InvariantStopping.Policy.HittingTime — Function

HittingTime

InvariantStopping.Policy.DeterministicTime — Function

DeterministicTime

Schedule

InvariantStopping.Scheduler.Schedule — Type

Schedule

Meant to be as easy as possible for the user to define.

LoweredSchedule

InvariantStopping.Scheduler.LoweredSchedule — Type

LoweredSchedule

UnderlyingModel

InvariantStopping.Transition.BrownianMotion — Type

BrownianMotion

InvariantStopping.Transition.GeometricBrownianMotion — Type

GeometricBrownianMotion <: UnderlyingModel

A geometric Brownian motion, specified by its rate, standard deviation sigma and dividend.

julia> underlying_model = GeometricBrownianMotion(3.1,2.0,0.0)

Sample

InvariantStopping.Sampler.Sample — Type

Sample

Each of the later is to be interpreted as happening at a later point in time.

LoweredSample

InvariantStopping.Sampler.LoweredSample — Type

LoweredSample

This type stores the basic unit of sampling events.